Changed to xib
Changed CharacterItemCell to xib file
