+++ Marvel Encyclopedia/View/Components/CharacterItem/CharacterItem.xib
@@ -1,0 +1,70 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="22505" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
+    <device id="retina6_12" orientation="portrait" appearance="light"/>
+    <dependencies>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="22504"/>
+        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <tableViewCell contentMode="scaleToFill" selectionStyle="default" indentationWidth="10" rowHeight="128" id="KGk-i7-Jjw" customClass="CharacterItem" customModule="Marvel_Encyclopedia" customModuleProvider="target">
+            <rect key="frame" x="0.0" y="0.0" width="351" height="128"/>
+            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+            <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" tableViewCell="KGk-i7-Jjw" id="H2p-sc-9uM">
+                <rect key="frame" x="0.0" y="0.0" width="351" height="128"/>
+                <autoresizingMask key="autoresizingMask"/>
+                <subviews>
+                    <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="ZhC-kH-oga">
+                        <rect key="frame" x="0.0" y="0.0" width="272.66666666666669" height="332.33333333333331"/>
+                        <subviews>
+                            <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleToFill" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="aa4-P8-Fji">
+                                <rect key="frame" x="10" y="5" width="257.66666666666669" height="322.33333333333331"/>
+                            </imageView>
+                        </subviews>
+                        <constraints>
+                            <constraint firstAttribute="trailing" secondItem="aa4-P8-Fji" secondAttribute="trailing" constant="5" id="Gca-k0-rAC"/>
+                            <constraint firstItem="aa4-P8-Fji" firstAttribute="leading" secondItem="ZhC-kH-oga" secondAttribute="leading" constant="10" id="RwN-gA-ut5"/>
+                            <constraint firstItem="aa4-P8-Fji" firstAttribute="top" secondItem="ZhC-kH-oga" secondAttribute="top" constant="5" id="XJH-Cz-cSc"/>
+                            <constraint firstAttribute="bottom" secondItem="aa4-P8-Fji" secondAttribute="bottom" constant="5" id="ltf-T8-8Sp"/>
+                            <constraint firstAttribute="width" secondItem="ZhC-kH-oga" secondAttribute="height" multiplier="105:128" id="z5p-3w-f6t"/>
+                        </constraints>
+                    </view>
+                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="dAY-or-aWF">
+                        <rect key="frame" x="279.66666666666669" y="9" width="51.333333333333314" height="24"/>
+                        <fontDescription key="fontDescription" type="boldSystem" pointSize="20"/>
+                        <nil key="textColor"/>
+                        <nil key="highlightedColor"/>
+                    </label>
+                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" text="Label" textAlignment="justified" lineBreakMode="tailTruncation" numberOfLines="5" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="5gW-sp-s1G">
+                        <rect key="frame" x="279.66666666666669" y="41" width="51.333333333333314" height="271.33333333333331"/>
+                        <fontDescription key="fontDescription" type="system" pointSize="14"/>
+                        <nil key="textColor"/>
+                        <nil key="highlightedColor"/>
+                    </label>
+                </subviews>
+                <constraints>
+                    <constraint firstAttribute="trailing" secondItem="dAY-or-aWF" secondAttribute="trailing" constant="20" symbolic="YES" id="1xB-Os-hDR"/>
+                    <constraint firstItem="ZhC-kH-oga" firstAttribute="leading" secondItem="H2p-sc-9uM" secondAttribute="leading" id="4Hn-Vv-Ga0"/>
+                    <constraint firstItem="5gW-sp-s1G" firstAttribute="leading" secondItem="ZhC-kH-oga" secondAttribute="trailing" constant="7" id="Ay0-d6-Q1O"/>
+                    <constraint firstItem="5gW-sp-s1G" firstAttribute="top" secondItem="dAY-or-aWF" secondAttribute="bottom" constant="8" symbolic="YES" id="D3v-m2-HeU"/>
+                    <constraint firstAttribute="bottom" secondItem="5gW-sp-s1G" secondAttribute="bottom" constant="20" symbolic="YES" id="Rer-fo-vtM"/>
+                    <constraint firstItem="dAY-or-aWF" firstAttribute="top" secondItem="H2p-sc-9uM" secondAttribute="top" constant="9" id="Ryt-SX-1ez"/>
+                    <constraint firstAttribute="trailing" secondItem="5gW-sp-s1G" secondAttribute="trailing" constant="20" symbolic="YES" id="WaO-l9-1go"/>
+                    <constraint firstItem="ZhC-kH-oga" firstAttribute="width" secondItem="H2p-sc-9uM" secondAttribute="width" multiplier="0.3" id="dJb-gC-cxC"/>
+                    <constraint firstItem="ZhC-kH-oga" firstAttribute="top" secondItem="H2p-sc-9uM" secondAttribute="top" id="dhE-La-iD0"/>
+                    <constraint firstItem="dAY-or-aWF" firstAttribute="leading" secondItem="ZhC-kH-oga" secondAttribute="trailing" constant="7" id="qfE-H2-7fX"/>
+                    <constraint firstItem="ZhC-kH-oga" firstAttribute="height" secondItem="H2p-sc-9uM" secondAttribute="height" id="t89-S8-eeH"/>
+                </constraints>
+            </tableViewCellContentView>
+            <viewLayoutGuide key="safeArea" id="njF-e1-oar"/>
+            <connections>
+                <outlet property="descLabel" destination="5gW-sp-s1G" id="ytf-Yv-IGA"/>
+                <outlet property="nameLabel" destination="dAY-or-aWF" id="j4L-3L-bpp"/>
+                <outlet property="thumbnail" destination="aa4-P8-Fji" id="pg9-SQ-Dsw"/>
+            </connections>
+            <point key="canvasLocation" x="124.42748091603053" y="64.788732394366207"/>
+        </tableViewCell>
+    </objects>
+</document>
